two unit tests
